safari-600.7-branch/LayoutTests/platform/ios-sim/fast/events/spatial-navigation/snav-iframe-with-offscreen-focusable-element-expected.txt

a

e
FAIL document.activeElement.getAttribute("id") should be 1. Was start.
FAIL document.activeElement.getAttribute("id") should be 1. Was start.
FAIL document.activeElement.getAttribute("id") should be 1. Was start.
FAIL document.activeElement.getAttribute("id") should be 2. Was start.
FAIL document.activeElement.getAttribute("id") should be end. Was start.
FAIL document.activeElement.getAttribute("id") should be 2. Was start.
FAIL document.activeElement.getAttribute("id") should be 2. Was start.
FAIL document.activeElement.getAttribute("id") should be 2. Was start.
FAIL document.activeElement.getAttribute("id") should be 1. Was start.
PASS document.activeElement.getAttribute("id") is "start"
